Move 'initialize_raw_file_iterator()' into preprocessing file
